updated code and data structure so that working parrallell will be easier, read the readme for explanation, you can also text me
