--- .vscode/launch.json
@@ -1,16 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.1",
-  "configurations": [
-    {
-      "name": "Python: Debug test",
-      "type": "python",
-      "request": "launch",
-      "program": "${workspaceDir}/icp/src/main.py",
-      "console": "integratedTerminal",
-      "justMyCode": true
-    }
-  ]
-}
