--- dbow/lib/fbow/src/fbow_exports.h
@@ -1,51 +1,0 @@
-/*****************************
-Copyright 2014 Rafael Muñoz Salinas. All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification, are
-permitted provided that the following conditions are met:
-
-   1. Redistributions of source code must retain the above copyright notice, this list of
-      conditions and the following disclaimer.
-
-   2. Redistributions in binary form must reproduce the above copyright notice, this list
-      of conditions and the following disclaimer in the documentation and/or other materials
-      provided with the distribution.
-
-THIS SOFTWARE IS PROVIDED BY Rafael Muñoz Salinas ''AS IS'' AND ANY EXPRESS OR IMPLIED
-WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
-FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL Rafael Muñoz Salinas OR
-CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
-ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
-NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
-ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-The views and conclusions contained in the software and documentation are those of the
-authors and should not be interpreted as representing official policies, either expressed
-or implied, of Rafael Muñoz Salinas.
-********************************/
- 
-
-
-#ifndef __FBOW_CORE_TYPES_H__
-#define __FBOW_CORE_TYPES_H__
- 
-#if !defined _CRT_SECURE_NO_DEPRECATE && _MSC_VER > 1300
-#define _CRT_SECURE_NO_DEPRECATE /* to avoid multiple Visual Studio 2005 warnings */
-#endif
-
-#if (defined WIN32 || defined _WIN32 || defined WINCE)   && defined FBOW_DSO_EXPORTS
-    #define FBOW_API __declspec(dllexport)
-    #pragma warning ( disable : 4251 ) //disable warning to templates with dll linkage.
-    #pragma warning ( disable : 4290 ) //disable warning due to exception specifications.
-    #pragma warning ( disable : 4996 ) //disable warning regarding unsafe vsprintf.
-    #pragma warning ( disable : 4244 ) //disable warning convesions with lost of data.
-
-#else
-    #define FBOW_API
-#endif
-
-
-#define  FBOW_VERSION "${PROJECT_VERSION}"
-#endif
