--- dbow/lib/fbow/tests/README_DBOW2
@@ -1,1 +1,0 @@
-DBOW2 version extracted from ORBSLAM2 and removed DLIB dependency
