--- dbow/lib/fbow/tests/dbow2/FORB.cpp
@@ -1,28 +1,0 @@
-/**
- * File: FORB.cpp
- * Date: June 2012
- * Author: Dorian Galvez-Lopez
- * Description: functions for ORB descriptors
- * License: see the LICENSE.txt file
- *
- * Distance function has been modified 
- *
- */
-
- 
-#include <vector>
-#include <string>
-#include <sstream>
-#include <stdint-gcc.h>
-
-#include "FORB.h"
-
-using namespace std;
-
-namespace DBoW2 {
-
-// --------------------------------------------------------------------------
-
-} // namespace DBoW2
-
-
